+++ src/main/java/com/beyond/easycheck/user/application/service/UserReadUseCase.java
@@ -1,4 +1,4 @@
-package com.beyond.easycheck.user.application.service.user;
+package com.beyond.easycheck.user.application.service;
 
 import com.beyond.easycheck.user.infrastructure.persistence.mariadb.entity.user.UserEntity;
 import lombok.Builder;
