+++ src/main/java/com/beyond/easycheck/user/ui/controller/UserController.java
@@ -1,9 +1,9 @@
 package com.beyond.easycheck.user.ui.controller;
 
-import com.beyond.easycheck.user.application.service.admin.AdminOperationUseCase;
-import com.beyond.easycheck.user.application.service.admin.AdminOperationUseCase.UserStatusUpdateCommand;
-import com.beyond.easycheck.user.application.service.user.UserOperationUseCase;
-import com.beyond.easycheck.user.application.service.user.UserReadUseCase;
+import com.beyond.easycheck.admin.application.service.AdminOperationUseCase;
+import com.beyond.easycheck.admin.application.service.AdminOperationUseCase.UserStatusUpdateCommand;
+import com.beyond.easycheck.user.application.service.UserOperationUseCase;
+import com.beyond.easycheck.user.application.service.UserReadUseCase;
 import com.beyond.easycheck.user.ui.requestbody.ChangePasswordRequest;
 import com.beyond.easycheck.user.ui.requestbody.UserLoginRequest;
 import com.beyond.easycheck.user.ui.requestbody.UserRegisterRequest;
@@ -23,10 +23,10 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import static com.beyond.easycheck.user.application.service.user.UserOperationUseCase.*;
-import static com.beyond.easycheck.user.application.service.user.UserReadUseCase.*;
-import static com.beyond.easycheck.user.application.service.user.UserReadUseCase.FindJwtResult;
-import static com.beyond.easycheck.user.application.service.user.UserReadUseCase.FindUserResult;
+import static com.beyond.easycheck.user.application.service.UserOperationUseCase.*;
+import static com.beyond.easycheck.user.application.service.UserReadUseCase.*;
+import static com.beyond.easycheck.user.application.service.UserReadUseCase.FindJwtResult;
+import static com.beyond.easycheck.user.application.service.UserReadUseCase.FindUserResult;
 
 @Slf4j
 @RestController
